(DOES NOT CURRENTLY COMPILE) Started on backend of mentoring relationship, changed User ID to be of type long, added a test related to Mentees

+++ api/baytree_mentoring/src/main/java/com/baytree_mentoring/baytree_mentoring/models/User.java
@@ -12,7 +12,7 @@ import javax.persistence.Id;
 @NoArgsConstructor
 public class User {
 
-    public User(int viewsId, String firstName, String lastName, String name, String email,
+    public User(long viewsId, String firstName, String lastName, String name, String email,
                 String status, String startDate, String endDate, String phoneNumber,
                 String ethnicity, String address, String role) {
         this.viewsId = viewsId;
@@ -30,7 +30,7 @@ public class User {
     }
 
     @Id
-    private int viewsId;
+    private long viewsId;
 
     String firstName;
